backend: Refactor config classes
Remove the configs module by moving the various configs to where they
are actually used. This simplifies the import dependencies within Kale
and allows to re-use common functions and constants, when needed.
