+++ backend/kale/processors/__init__.py
@@ -12,4 +12,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from .nbprocessor import NotebookProcessor
+from .nbprocessor import NotebookProcessor, NotebookConfig
