+++ backend/kale/tests/unit_tests/test_compiler.py
@@ -17,8 +17,7 @@ import pytest
 
 from unittest import mock
 
-from kale import Pipeline, Step, Compiler
-from kale.config import NotebookConfig
+from kale import Pipeline, Step, Compiler, NotebookConfig
 
 
 THIS_DIR = os.path.dirname(os.path.abspath(__file__))
